requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==4.2.17
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
drf-yasg==1.21.11
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
mercadopago==2.3.0
mysql-connector-python==9.5.0
packaging==25.0
PyMySQL==1.1.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.11.0
